+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [1.0.1] - 2025-12-21
 
 - Fix spec_helper/rails_helper path finding [Thanks @diego-aslz]
+- Add --prefork-require and --no-prefork-require CLI options for non-rails apps or rails setups where loading config/application.rb is not entirely safe
 
 ## [1.0.0] - 2025-12-21