+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## [1.0.2] - 2026-01-09
+- Fix --postfork-require options
+
 ## [1.0.1] - 2025-12-21
 
 - Fix spec_helper/rails_helper path finding [Thanks @diego-aslz]